requirements-ned.txt
requests
pyyaml
anthropic
youtube-transcript-api
google-api-python-client
